update internal date api
added function `localeDate()` warpping the `Date.toLocaleString()`
function.

this was added because our `date` metadata might be invalid, in which
case we should display the original string instead of showing
`undefined`.

the function was removed in a previous commit, and is brought back
because of convenience. (the implementation isn't quite understandable,
and is used in more than 1 place.)