add svelte preprocessor to automatically localize anchor href
This is a preprocessor that automatically wrap href's in ParaglideJS's localizeHref function.
It is brought up by Paraglide staff after ParaglideJS v2 no longer supporting doing it.
It does the function import, the content swapping and the variable inserting, and generally works fine for local paths.
However, it is also adding the locale to external links (e.g. `github.com/en/explor`), which could probably be fixed with the url strategy patterns. (It can not be detected by the preprocessor if the external url is stored in a variable.)
Also, it skips expressions with localizeHref, which does prevent localized paths to be wrapped, however my implementation is not strict enough that it may cause problems.

Generally, this preprocessor should work if other things are also well configured. The only concern is if there are any edge cases currently not handled by the code.